refactor!: Un-fork base typesetter and make a silex prototype typesetter
